% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_data}
\alias{load_data}
\title{Load the data from the Harvard Dataverse}
\usage{
load_data(file_name = NULL, file_note = NULL)
}
\arguments{
\item{file_name}{The exact file name for a particular dataset.One can find this information usign \verb{view_data().}}

\item{file_note}{The exact note for a particular dataset. One can find this information using \verb{view_data().}}
}
\value{
df A particular dataset
}
\description{
Load the data from the Harvard Dataverse
}
