% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_names.R
\name{select_names}
\alias{select_names}
\title{Select names}
\usage{
select_names(race, pct_correct_thres = 0.8, n_names = 5)
}
\arguments{
\item{race}{The race intended to be signaled by the names. The value should be either "Asian," "Hispanic," "Black," or "White."}

\item{pct_correct_thres}{The names' threshold rate of being correctly perceived as a particular race. The default value is 0.8}

\item{n_names}{The number of the names to be selected. The list is ordered by pct_correct. The default number is 5.}
}
\value{
The dataset that contains the list of the selected names. first = first name, last = last name, w.asian = westernized Asian name (1 = yes, 0 = no), name (full name), identity (the name's intended race),
pct_correct = the percentage of the name's intended race correctly perceived, avg_income = average perceived income level, avg_edu = average perceived education level, avg_citizenship_guess = the percentage of perceived citizenship status
}
\description{
Select names
}
