% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{view_data}
\alias{view_data}
\title{View the data from the Harvard Dataverse}
\usage{
view_data(name_type = NULL, name_target = NULL)
}
\arguments{
\item{name_type}{The type of the dataset. Either original or validated. The default value is NULL.}

\item{name_target}{The name of the target dataset. The default value is NULL.}
}
\value{
A dataframe that contains three columns: file name, type (either origianl or validated), and note
}
\description{
View the data from the Harvard Dataverse
}
