% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{load_data}
\alias{load_data}
\title{Load the data from the Harvard Dataverse}
\usage{
load_data(file_name = NULL, file_note = NULL)
}
\arguments{
\item{file_name}{The exact file name for a particular dataset. Please use \code{view_data()} to find this information.}

\item{file_note}{The exact note for a particular dataset. Please use \code{view_data()} to find this information.}
}
\value{
df A particular dataset
}
\description{
Load the data from the Harvard Dataverse
}
