% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_names.R
\name{select_names}
\alias{select_names}
\title{Select names}
\usage{
select_names(race, pct_correct_thres = 0.8, order_by_var = NULL, n_names = 5)
}
\arguments{
\item{race}{Character. The race intended to be signaled by the names. The value should be either "Asian," "Hispanic," "Black," or "White."}

\item{pct_correct_thres}{Numeric. The names' threshold rate of being correctly perceived as a particular race. The default value is 0.8}

\item{order_by_var}{Character. The variable to be used to order the list of the names to be extracted. The default variable is NULL (non-selected). In this case, the names will be chosen randomly. The other options are pct_correct = the percentage of the name's intended race correctly perceived, avg_income = average perceived income level, avg_education = average perceived education level, and avg_citizenship_guess = the percentage of perceived citizenship status.}

\item{n_names}{Numeric. The number of the names to be selected. The default number is 5.}
}
\value{
The dataset that contains the list of the selected names. first = first name, last = last name, w.asian = westernized Asian name (1 = yes, 0 = no), name (full name), identity (the name's intended race),
pct_correct = the percentage of the name's intended race correctly perceived, avg_income = average perceived income level, avg_education = average perceived education level, avg_citizenship_guess = the percentage of perceived citizenship status
}
\description{
Select names
}
